requirements.txt
aiofiles==23.2.1
aioshutil==1.3
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apify==1.4.1
apify-client==1.6.1
apify-shared==1.1.0
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
distro==1.8.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
openai==1.6.1
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.7
pyarrow==14.0.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
streamlit==1.29.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
websockets==12.0
zipp==3.17.0
